+++ CHANGELOG.md
@@ -18,3 +18,4 @@
 - Poison Proof Gene now uses the new NeoForge Poison damage, rather than checking for magic damage while you have poison effect
 - Renamed The Cure to Panacea
 - The Coal Generator now allows extracting items from any side (in case of fuels like Lava Buckets, which leave an empty Bucket behind)
+- Machines now keep their progress if they run out of power